Fixed some warnings from ScalaStyle
