+++ project/plugins.sbt
@@ -1,2 +1,3 @@
 logLevel := Level.Warn
 addSbtPlugin("org.scoverage" % "sbt-scoverage" % "1.5.0")
+addSbtPlugin("org.scalastyle" %% "scalastyle-sbt-plugin" % "0.8.0")
